Product API 구현

+++ src/main/java/smu/likelion/jikchon/service/ProductService.java
@@ -4,15 +4,19 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import smu.likelion.jikchon.base.PageResult;
+import smu.likelion.jikchon.base.SubCategory;
 import smu.likelion.jikchon.domain.member.Member;
 import smu.likelion.jikchon.domain.Product;
 import smu.likelion.jikchon.dto.product.ProductRequestDto;
 import smu.likelion.jikchon.dto.product.ProductReturnDto;
+import smu.likelion.jikchon.exception.CustomException;
 import smu.likelion.jikchon.exception.CustomNotFoundException;
 import smu.likelion.jikchon.exception.ErrorCode;
 import smu.likelion.jikchon.repository.MemberRepository;
 import smu.likelion.jikchon.repository.ProductRepository;
 import org.springframework.data.domain.Pageable;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -46,15 +50,18 @@ public class ProductService{
         Member member = memberRepository.findById(loginService.getLoginMemberId()).orElseThrow(()->{
             throw new CustomNotFoundException(ErrorCode.NOT_FOUND_MEMBER);
         });
-        productRepository.save(productRequestDto.toEntity(member));
+        productRepository.save(productRequestDto.toEntity());
     }
     //프로덕트 수정
     public void update(Long id, ProductRequestDto productRequestDto) {
         Product productData=productRepository.findById(id).orElseThrow(()->{
             throw new CustomNotFoundException(ErrorCode.NOT_FOUND);
         });
+        if(!productData.getMember().getId().equals(loginService.getLoginMemberId())){
+            throw new CustomException(ErrorCode.FORBIDDEN);
+        }
         productData.setProductName(productRequestDto.getProductName());
-//        productData.setCategory(productRequestDto.getCategory());
+        productData.setSubCategory(SubCategory.fromDescription(productRequestDto.getSubCategory()));
         productData.setPrice(productRequestDto.getPrice());
         productData.setQuantity(productRequestDto.getQuantity());
         productData.setIntro(productRequestDto.getIntro());
@@ -65,6 +72,9 @@ public class ProductService{
         Product productData=productRepository.findById(id).orElseThrow(()->{
             throw new CustomNotFoundException(ErrorCode.NOT_FOUND);
         });
+        if(!productData.getMember().getId().equals(loginService.getLoginMemberId())){
+            throw new CustomException(ErrorCode.FORBIDDEN);
+        }
         productRepository.delete(productData);
     }
 }
